requirements.txt
flask
cryptography
requests
gunicorn
pymupdf
python-docx